requirements.txt
anyio==3.6.1
attrs==22.1.0
boto3==1.24.82
botocore==1.27.82
click==8.1.3
fastapi==0.85.0
fastapi-jwt-auth==0.5.0
greenlet==1.1.3
h11==0.14.0
idna==3.4
iniconfig==1.1.1
jmespath==1.0.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pydantic==1.10.2
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-slugify==6.1.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
